requirements/common.txt
Django==1.6.2
South==0.8.4
psycopg2==2.5.2
Unipath
django-registration==1.0
cloudinary
